+++ sharding-core/sharding-core-execute/src/main/java/org/apache/shardingsphere/core/execute/sql/StatementExecuteUnit.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.execute;
+package org.apache.shardingsphere.core.execute.sql;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
