+++ sharding-core/sharding-core-execute/src/test/java/org/apache/shardingsphere/core/execute/sql/prepare/SQLExecutePrepareTemplateTest.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.core.execute.sql.prepare;
 
 import org.apache.shardingsphere.core.constant.ConnectionMode;
 import org.apache.shardingsphere.core.execute.ShardingExecuteGroup;
-import org.apache.shardingsphere.core.execute.StatementExecuteUnit;
+import org.apache.shardingsphere.core.execute.sql.StatementExecuteUnit;
 import org.apache.shardingsphere.core.route.RouteUnit;
 import org.apache.shardingsphere.core.route.SQLUnit;
 import org.junit.Test;
